Fixed bugs and warnings

+++ src/Chirp.Web/Pages/InformationPageView.cshtml.cs
@@ -17,7 +17,7 @@ public class InformationPageView : PageModel
     public AuthorDTO? Author { get; set; } = null;  //Tracker til author navn
     public int CurrentCheeps { get; set; } = 0;
 
-    public IEnumerable<AuthorDTO> Following { get; set; } = new List<AuthorDTO>();
+    public IEnumerable<AuthorDTO?> Following { get; set; } = new List<AuthorDTO>();
 
     public InformationPageView(ICheepService service, IAuthorService authorService, IIdentityUserService identityService)
     {
@@ -30,7 +30,10 @@ public class InformationPageView : PageModel
     {
         if (_identityService.IsSignedIn(User))
         {
-            AuthorDTO CurrentAuthor = await _identityService.GetCurrentIdentityAuthor(User);
+            AuthorDTO? CurrentAuthor = await _identityService.GetCurrentIdentityAuthor(User);
+
+            if (CurrentAuthor == null)
+                return Page();
 
             await _cheepService.DeleteAllCheepsAsync(CurrentAuthor.Id);
             _authorService.RemoveAllFollowers(CurrentAuthor.Id);
@@ -62,7 +65,7 @@ public class InformationPageView : PageModel
     public async Task<ActionResult> OnPostUnfollow(string followeeId)
     {
 
-        AuthorDTO currentAuthor = await _identityService.GetCurrentIdentityAuthor(User);
+        AuthorDTO? currentAuthor = await _identityService.GetCurrentIdentityAuthor(User);
 
         if (currentAuthor == null)
         {
@@ -85,20 +88,10 @@ public class InformationPageView : PageModel
             return null;
         }
 
-        AuthorDTO author = await _identityService.GetCurrentIdentityAuthor(User);
+        AuthorDTO? author = await _identityService.GetCurrentIdentityAuthor(User);
 
         return author;
     }
 
-    public async Task<IEnumerable<AuthorDTO>> GetFollowers()
-    {
-        if (!_identityService.IsSignedIn(User))
-        {
-            return new List<AuthorDTO>();
-        }
 
-        AuthorDTO author = await _identityService.GetCurrentIdentityAuthor(User);
-
-        return _authorService.GetFollowers(author.Id);
-    }
 }
